update to 1D mesh, created Mesh1Dimension class
